refs #24580 [Training][Exam3][NguyenThaiTruong] Final Examination - Create API update order status. - Update API count today's order quantity. - Update getTodaySales to getSales(startDate, endDate). -  2.7 Orders  2.7.1  - Show all item which ordered  - Search name food/drink in all orders  - Sort : today ( order today), week ( order in this week), Month ( order in this month), All ( all orders)  2.7.2  + Today's sale : total money which sold today ( include completed, shipping, new)  + Total: all money which sold ( include completed, shipping, new)  2.8 Detail Order  - Show all item which ordered  - change status of order to Shipping, Completed or Cancel.  - Click X icon will close pop-up.

+++ JavaWeb_nttruong/src/main/java/com/foodbooking/entity/OrderStatusEnum.java
@@ -7,13 +7,13 @@ public enum OrderStatusEnum {
 	CANCELLED(3),
 	COMPLETED(4);
 	
-	private final int id;
+	private final long id;
 	
-	OrderStatusEnum(int id) {
+	OrderStatusEnum(long id) {
 		this.id = id;
 	}
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
